+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All Notable changes will be documented in this file
 
+## 1.1.0
+
+ - Added ability to update a token customer using client.update()
+ - Added SaveCustomer flag to Transactions to create a token when a transaction is completed.
+ - Added a debug flag to enable logging of errors
+
 ## 1.0.0
 
  - First release